Media:  - Fixed shaders multiplication order (again...)  - Renamed some shaders.
Core:
 - Started work on memory allocators (just API design tests).

Engine:
 - Reworked index buffers to be index size abstract.
 - Simplified the interface (removed useless getters).

Editor:
 - Started moving to a document-centric architecture.
 - Started work in generalizing the scene pane into a hierarchy pane.
 - Added a custom editor camera controller.

General:
 - Removed double usage from the code (all floats now).
 - Simplified code for speed reasons (like removing useless virtual destructors).

+++ inc/vapor/Subsystem.h
@@ -25,7 +25,7 @@ class CORE_API Subsystem
 public:
 
 	virtual ~Subsystem() { }
-	virtual void update( double delta ) { }
+	virtual void update( float delta ) { }
 
 protected:
