+++ inc/vapor/controllers/CameraController.h
@@ -25,7 +25,7 @@ class VAPOR_API CameraController : public Controller
 public:
 
 	// Updates this component.
-	virtual void _update( double delta ) = 0;
+	virtual void _update( float delta ) = 0;
 
 	// Gets/sets the look sensivity of the camera.
 	ACESSOR(LookSensivity, float, lookSensivity);
