+++ inc/vapor/physics/MeshShape.h
@@ -33,7 +33,7 @@ public:
 	virtual ~MeshShape();
 
 	// Updates the shape.
-	void update( double delta );
+	void update( float delta );
 
 	// Gets the internal Bullet shape.
 	btCollisionShape* const getBulletShape() const;
