+++ src/Core/Profiler.cpp
@@ -22,8 +22,8 @@ Profiler::Profiler(const std::string& name)
 
 Profiler::~Profiler()
 {
-	double time = timer.getElapsedTime();
-	double ms = time * 1000;
+	float time = timer.getElapsedTime();
+	float ms = time * 1000;
 
 	Log::debug( "%s: took %lu ms (%.5lf)", name.c_str(), (ulong) ms, ms );
 }
